feat: oriented object implementation
add classes:
- browser
- navigator
- scraper
- study
