+++ src/scraper/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Oct 25 17:14:38 2020
+
+@author: jpuig
+"""
+
