+++ src/scraper/main.py
@@ -5,8 +5,6 @@ Created on Sun Oct 18 15:29:47 2020
 @author: jpuig
 """
 
-
-
 class Main:
  
   test = Scraper()
