+++ src/scraper/navigator.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Oct 25 17:03:08 2020
+
+@author: jpuig
+"""
+class NavigatorScraper:
+  def __init__(self, browser):          
+      self.browser = browser
+      
+  def get_next_link(self):
+      class_name = 'tr-next-link'
+      next_link = self.browser.find_element_by_class_name(class_name)
+      next_link.click()
+
+  def get_previous_link(self):
+      class_name = 'tr-prev-link'
+      previous_link = self.browser.find_element_by_class_name(class_name)
+      previous_link.click()
