+++ src/scraper/studies.csv
@@ -3,4 +3,3 @@ NCT04469647;COVID 19 Seroprevalence Amongst Healthcare Workers in JHAH;COVID 19 
 NCT04427254;Study of the Sars-Cov2 Neuroinvasiveness - COVID19;Study of the Sars-Cov2 Neuroinvasiveness - COVID19
 NCT04377685;Prediction of Clinical Course in COVID19 Patients;Prediction of Clinical Course in COVID19 Patients Using Unsupervised Classification Approaches of Clinical, Biological and the Multiparametric Signature of the Chest CT Scan Performed at Admission
 NCT04386668;Let It Out (LIO) and COVID19: Testing an Online Emotional Disclosure-based Intervention During the COVID19 Pandemic;Let It Out (LIO) and COVID19: a Randomised Controlled Trial of an Online Emotional Disclosure-based Intervention for Adults During the COVID19 Pandemic
-
