+++ src/scraper/study.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Oct 25 17:05:26 2020
+
+@author: jpuig
+"""
+
+class Study():    
+  def __init__(self):
+     pass
+     
+  def id(self):
+      return self.id
+
+  def title(self):
+      return self.title    
+
+  def official_title(self):
+      return self.official_title
